Stocks/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
curl_cffi==0.13.0
Flask==3.0.0
frozendict==2.4.7
html5lib==1.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==1.26.4
pandas==2.2.0
peewee==3.19.0
platformdirs==4.4.0
protobuf==6.33.4
pycparser==2.23
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.25
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.5
yfinance==1.2.0
zipp==3.23.0
